move locations to their own module
