+++ src/ast/mod.rs
@@ -1,63 +1,18 @@
-use std::fmt::{self, Debug, Display, Formatter};
+use std::fmt::{Debug, Display, Formatter, Result};
 
 use crate::ty::{Binding, Ty};
 
+mod location;
+
+pub use location::*;
+
 pub type Block<'a> = Vec<Located<Node<'a>>>;
-
-#[derive(Debug, Clone, Copy)]
-pub struct Location {
-    pub start: usize,
-    pub end: usize,
-}
-
-impl Location {
-    pub fn new(start: usize, end: usize) -> Self {
-        Location { start, end }
-    }
-}
-
-impl std::ops::Add for Location {
-    type Output = Self;
-    fn add(mut self, other: Self) -> Self {
-        self.end = other.end;
-        self
-    }
-}
-
-#[derive(Debug)]
-pub struct Located<T: Debug> {
-    pub content: T,
-    pub loc: Location,
-}
-
-impl<T: Debug> Located<T> {
-    pub fn new(content: T, loc: impl Into<Location>) -> Self {
-        Located {
-            content,
-            loc: loc.into(),
-        }
-    }
-}
-
-impl<T: Display + Debug> Display for Located<T> {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "{}", self.content)
-    }
-}
-
-impl<T: Eq + Debug> Eq for Located<T> {}
-
-impl<T: PartialEq + Debug> PartialEq for Located<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.content == other.content
-    }
-}
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 pub struct Name<'a>(pub &'a str);
 
 impl<'a> Display for Name<'a> {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut Formatter) -> Result {
         write!(f, "{}", self.0)
     }
 }
@@ -85,7 +40,7 @@ pub enum BinOp {
 }
 
 impl<'a> Display for BinOp {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut Formatter) -> Result {
         use BinOp::*;
         match self {
             Add => write!(f, "+"),
@@ -117,7 +72,7 @@ pub enum UnOp {
 }
 
 impl<'a> Display for UnOp {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut Formatter) -> Result {
         use UnOp::*;
         match self {
             Not => write!(f, "!"),
@@ -146,7 +101,7 @@ impl Into<Literal> for bool {
 }
 
 impl<'a> Display for Literal {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut Formatter) -> Result {
         use Literal::*;
         match self {
             Bool(b) => write!(f, "{}", b),
